% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_ace}
\alias{bt_ace}
\title{ACE Richness}
\usage{
bt_ace(x)
}
\arguments{
\item{x}{Community matrix.}
}
\value{
Numeric vector.
}
\description{
Estimates species richness using ACE.
}
\details{
Wrapper around vegan::estimateR().
}
\examples{
mat <- matrix(
c(
5,2,0,
4,1,3
),
nrow=2,
byrow=TRUE
)

bt_ace(mat)

}
