% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_adjusted_r_squared}
\alias{bt_adjusted_r_squared}
\title{Adjusted R-squared}
\usage{
bt_adjusted_r_squared(model)
}
\arguments{
\item{model}{A linear model.}
}
\value{
Numeric.
}
\description{
Returns adjusted R-squared.
}
\examples{
birds <- data.frame(
  y = c(20,25,18,30,27),
  x = c(12,14,11,16,15)
)

model <- lm(y ~ x, data = birds)

bt_adjusted_r_squared(model)

}
