% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_aic_weights}
\alias{bt_aic_weights}
\title{Akaike Weights}
\usage{
bt_aic_weights(...)
}
\arguments{
\item{...}{Two or more fitted model objects.}
}
\value{
A data frame.
}
\description{
Calculates Akaike weights from multiple fitted models.
}
\examples{
birds <- data.frame(
  y = c(10,12,13,16,18),
  x1 = c(2,3,4,5,6),
  x2 = c(5,4,6,7,8)
)

m1 <- lm(y ~ x1, data = birds)
m2 <- lm(y ~ x2, data = birds)

bt_aic_weights(m1,m2)

}
