% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_anova}
\alias{bt_anova}
\title{One-way ANOVA}
\usage{
bt_anova(data, response, group)
}
\arguments{
\item{data}{A data frame.}

\item{response}{Numeric variable.}

\item{group}{Grouping variable.}
}
\value{
anova object.
}
\description{
Performs one-way ANOVA.
}
\examples{
birds <- data.frame(
  mass=c(20,22,25,28,30,35),
  diet=c("A","A","B","B","C","C")
)

bt_anova(
  birds,
  "mass",
  "diet"
)

}
