% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_anova_models}
\alias{bt_anova_models}
\title{Compare Models Using ANOVA}
\usage{
bt_anova_models(...)
}
\arguments{
\item{...}{Two or more fitted model objects.}
}
\value{
An ANOVA table.
}
\description{
Compares two or more fitted models using analysis of variance.
}
\details{
This function is a wrapper around \code{stats::anova()} and is
primarily intended for comparing nested linear models.
}
\examples{
birds <- data.frame(
  y = c(10,12,13,16,18),
  x1 = c(2,3,4,5,6),
  x2 = c(5,4,6,7,8)
)

m1 <- lm(y ~ x1, data = birds)
m2 <- lm(y ~ x1 + x2, data = birds)

bt_anova_models(m1, m2)

}
