% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_best_model}
\alias{bt_best_model}
\title{Best Model}
\usage{
bt_best_model(...)
}
\arguments{
\item{...}{Two or more fitted model objects.}
}
\value{
A fitted model object.
}
\description{
Returns the model with the lowest AIC.
}
\examples{
birds <- data.frame(
  y = c(10,12,13,16,18),
  x1 = c(2,3,4,5,6),
  x2 = c(5,4,6,7,8)
)

m1 <- lm(y ~ x1, data = birds)
m2 <- lm(y ~ x2, data = birds)

best <- bt_best_model(m1, m2)

}
