% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_beta_diversity}
\alias{bt_beta_diversity}
\title{Beta Diversity}
\usage{
bt_beta_diversity(x, method = "bray")
}
\arguments{
\item{x}{Community matrix (sites x species).}

\item{method}{Distance method.}
}
\value{
A dist object.
}
\description{
Calculates pairwise beta diversity distances between sampling units.
}
\details{
Wrapper around vegan::vegdist().
}
\examples{
mat <- matrix(
c(
5,2,0,
4,1,3,
0,5,2,
2,1,6
),
nrow = 4,
byrow = TRUE
)

bt_beta_diversity(mat)

}
