% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_bic}
\alias{bt_bic}
\title{Bayesian Information Criterion}
\usage{
bt_bic(model)
}
\arguments{
\item{model}{A fitted model.}
}
\value{
Numeric.
}
\description{
Returns the Bayesian Information Criterion (BIC).
}
\examples{
birds <- data.frame(
  y = c(20,25,18,30,27),
  x = c(12,14,11,16,15)
)

model <- lm(y ~ x, data = birds)

bt_bic(model)

}
