% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_bray_curtis}
\alias{bt_bray_curtis}
\title{Bray-Curtis Distance}
\usage{
bt_bray_curtis(x)
}
\arguments{
\item{x}{Community matrix.}
}
\value{
A dist object.
}
\description{
Calculates Bray-Curtis dissimilarity.
}
\details{
Wrapper around vegan::vegdist().
}
\examples{
mat <- matrix(
c(
5,2,0,
4,1,3,
0,5,2,
2,1,6
),
nrow = 4,
byrow = TRUE
)

bt_bray_curtis(mat)

}
