% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_chao1}
\alias{bt_chao1}
\title{Chao1 Richness}
\usage{
bt_chao1(x)
}
\arguments{
\item{x}{Community matrix.}
}
\value{
Numeric vector.
}
\description{
Estimates species richness using Chao1.
}
\details{
Wrapper around vegan::estimateR().
}
\examples{
mat <- matrix(
c(
5,2,0,
4,1,3
),
nrow=2,
byrow=TRUE
)

bt_chao1(mat)

}
