% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bt_character_columns}
\alias{bt_character_columns}
\title{Character Columns}
\usage{
bt_character_columns(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Character vector.
}
\description{
Returns the names of character columns.
}
\examples{
birds <- data.frame(
  species = c("A","B"),
  diet = c("Omnivore","Carnivore"),
  stringsAsFactors = FALSE
)

bt_character_columns(birds)

}
