% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bt_character_summary}
\alias{bt_character_summary}
\title{Character Summary}
\usage{
bt_character_summary(data, column)
}
\arguments{
\item{data}{A data frame.}

\item{column}{Character column.}
}
\value{
List.
}
\description{
Returns summary information for a character or factor column.
}
\examples{
birds <- data.frame(
  diet = c(
    "Insectivore",
    "Omnivore",
    "Omnivore",
    "Carnivore"
  ),
  stringsAsFactors = FALSE
)

bt_character_summary(birds, "diet")

}
