% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bt_check_columns}
\alias{bt_check_columns}
\title{Check Required Columns}
\usage{
bt_check_columns(data, required)
}
\arguments{
\item{data}{Data frame.}

\item{required}{Character vector.}
}
\value{
Invisibly returns TRUE.
}
\description{
Internal helper that checks required columns exist.
}
\keyword{internal}
