% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bt_check_community}
\alias{bt_check_community}
\title{Check Community Matrix}
\usage{
bt_check_community(data)
}
\arguments{
\item{data}{Object supplied by the user.}
}
\value{
Invisibly returns TRUE.
}
\description{
Internal helper that verifies the input is a community matrix.
}
\keyword{internal}
