% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_check_homoscedasticity}
\alias{bt_check_homoscedasticity}
\title{Check Homoscedasticity}
\usage{
bt_check_homoscedasticity(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
A list.
}
\description{
Computes the correlation between fitted values and squared residuals.
}
\details{
Small correlations indicate more homogeneous variance.
}
\examples{
birds <- data.frame(
  y = c(20,25,18,30,27,24),
  x = c(12,14,11,16,15,13)
)

model <- lm(y ~ x, data = birds)

bt_check_homoscedasticity(model)

}
