% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_check_normality}
\alias{bt_check_normality}
\title{Check Normality}
\usage{
bt_check_normality(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
An htest object.
}
\description{
Performs the Shapiro-Wilk test on model residuals.
}
\examples{
birds <- data.frame(
  y = c(20,25,18,30,27,24),
  x = c(12,14,11,16,15,13)
)

model <- lm(y ~ x, data = birds)

bt_check_normality(model)

}
