% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_check_vif}
\alias{bt_check_vif}
\title{Check Multicollinearity}
\usage{
bt_check_vif(model)
}
\arguments{
\item{model}{A fitted linear model.}
}
\value{
A named numeric vector.
}
\description{
Computes Variance Inflation Factors (VIF) for predictors in
a linear model.
}
\details{
Values greater than 5 may indicate problematic
multicollinearity.
}
\examples{
birds <- data.frame(
  y  = c(20,25,18,30,27,24,29,31),
  x1 = c(12,14,11,16,15,13,17,18),
  x2 = c(8,9,7,10,9,8,11,12)
)

model <- lm(y ~ x1 + x2, data = birds)

bt_check_vif(model)

}
