% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_chisq_test}
\alias{bt_chisq_test}
\title{Chi-square Test}
\usage{
bt_chisq_test(data, row, column)
}
\arguments{
\item{data}{A data frame.}

\item{row}{Row variable.}

\item{column}{Column variable.}
}
\value{
htest object.
}
\description{
Performs a Chi-square test between two categorical variables.
}
\examples{
birds <- data.frame(
  diet=c("A","A","B","B"),
  status=c("LC","VU","LC","LC")
)

bt_chisq_test(
  birds,
  "diet",
  "status"
)

}
