% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{bt_clean_data}
\alias{bt_clean_data}
\title{Clean Bird Dataset}
\usage{
bt_clean_data(data, replace_na = FALSE, na_value = "Unknown")
}
\arguments{
\item{data}{A data frame.}

\item{replace_na}{Logical. Replace NA values?}

\item{na_value}{Value used when replacing missing values.}
}
\value{
A cleaned data frame.
}
\description{
Cleans bird ecological datasets by:
}
\details{
\itemize{
\item removing empty rows
\item trimming whitespace
\item converting empty strings to NA
\item optionally replacing NA values
}
}
\examples{
birds <- data.frame(
  species = c(" Robin", "Sparrow ", ""),
  mass = c(20, NA, NA),
  stringsAsFactors = FALSE
)

birds <- bt_clean_data(birds)

}
