% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{bt_clean_names}
\alias{bt_clean_names}
\title{Clean Column Names}
\usage{
bt_clean_names(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Data frame.
}
\description{
Standardizes column names to snake_case.
}
