% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{bt_clean_strings}
\alias{bt_clean_strings}
\title{Clean Character Strings}
\usage{
bt_clean_strings(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Cleaned data frame.
}
\description{
Removes leading/trailing spaces and converts empty strings to NA.
}
