% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bt_column_classes}
\alias{bt_column_classes}
\title{Column Classes}
\usage{
bt_column_classes(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Data frame.
}
\description{
Returns the class of every column in the dataset.
}
\examples{
birds <- data.frame(
  species = c("A","B"),
  mass = c(20,25),
  stringsAsFactors = FALSE
)

bt_column_classes(birds)

}
