% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bt_column_summary}
\alias{bt_column_summary}
\title{Column Summary}
\usage{
bt_column_summary(data, column)
}
\arguments{
\item{data}{A data frame.}

\item{column}{Column name.}
}
\value{
List.
}
\description{
Returns a summary of a single column.
}
\examples{
birds <- data.frame(
  species = c("A","B","C","A"),
  stringsAsFactors = FALSE
)

bt_column_summary(birds, "species")

}
