% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_compare_models}
\alias{bt_compare_models}
\title{Compare Multiple Models}
\usage{
bt_compare_models(...)
}
\arguments{
\item{...}{Two or more fitted model objects.}
}
\value{
A data frame.
}
\description{
Compares fitted models using common model selection statistics.
}
\details{
Returns a summary table containing:
\itemize{
\item AIC
\item BIC
\item Log-Likelihood
\item Deviance
\item Residual Degrees of Freedom
}

Models are automatically ranked by AIC.
}
\examples{
birds <- data.frame(
  y=c(10,12,13,16,18),
  x1=c(2,3,4,5,6),
  x2=c(5,4,6,7,8)
)

m1 <- lm(y~x1,birds)
m2 <- lm(y~x2,birds)

bt_compare_models(m1,m2)

}
