% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_contingency_table}
\alias{bt_contingency_table}
\title{Contingency Table}
\usage{
bt_contingency_table(data, row, column)
}
\arguments{
\item{data}{A data frame.}

\item{row}{Row variable.}

\item{column}{Column variable.}
}
\value{
BirdTraitsContingency object.
}
\description{
Returns counts and percentages together.
}
\examples{
birds <- data.frame(
  diet = c(
    "Insectivore",
    "Omnivore",
    "Carnivore",
    "Insectivore"
  ),
  status = c(
    "LC",
    "LC",
    "VU",
    "NT"
  )
)

bt_contingency_table(
  birds,
  "diet",
  "status"
)

}
