% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_cooks_distance}
\alias{bt_cooks_distance}
\title{Cook's Distance}
\usage{
bt_cooks_distance(model)
}
\arguments{
\item{model}{A fitted regression model.}
}
\value{
Numeric vector.
}
\description{
Calculates Cook's Distance for each observation.
}
\examples{
model <- lm(mpg ~ wt, data = mtcars)
bt_cooks_distance(model)

}
