% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_correlation}
\alias{bt_correlation}
\title{Correlation}
\usage{
bt_correlation(data, x, y, method = "pearson")
}
\arguments{
\item{data}{A data frame.}

\item{x}{First numeric column.}

\item{y}{Second numeric column.}

\item{method}{Correlation method. One of
"pearson", "spearman", or "kendall".}
}
\value{
Numeric.
}
\description{
Calculates the correlation between two numeric variables.
}
\examples{
birds <- data.frame(
  mass = c(20,25,18,30,27),
  wing = c(12,14,11,16,15)
)

bt_correlation(
  birds,
  "mass",
  "wing"
)

}
