% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_correlation_matrix}
\alias{bt_correlation_matrix}
\title{Correlation Matrix}
\usage{
bt_correlation_matrix(data, method = "pearson")
}
\arguments{
\item{data}{A data frame.}

\item{method}{Correlation method.}
}
\value{
Matrix.
}
\description{
Calculates a correlation matrix for all numeric variables.
}
\examples{
birds <- data.frame(
  mass = c(20,25,18,30),
  wing = c(11,13,10,15),
  tail = c(7,8,6,9)
)

bt_correlation_matrix(birds)

}
