% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_count_families}
\alias{bt_count_families}
\title{Count Families}
\usage{
bt_count_families(data, family_col = "family")
}
\arguments{
\item{data}{Data frame.}

\item{family_col}{Family column.}
}
\value{
Integer.
}
\description{
Counts unique families.
}
