% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_count_orders}
\alias{bt_count_orders}
\title{Count Orders}
\usage{
bt_count_orders(data, order_col = "order")
}
\arguments{
\item{data}{Data frame.}

\item{order_col}{Order column.}
}
\value{
Integer.
}
\description{
Counts unique orders.
}
