% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_count_species}
\alias{bt_count_species}
\title{Count Species}
\usage{
bt_count_species(data, species_col = "species")
}
\arguments{
\item{data}{Data frame.}

\item{species_col}{Species column.}
}
\value{
Integer.
}
\description{
Counts unique species.
}
\examples{
birds <- data.frame(
  species = c("A","B","C","A")
)

bt_count_species(birds)

}
