% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_count_variables}
\alias{bt_count_variables}
\title{Count Variables}
\usage{
bt_count_variables(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Integer.
}
\description{
Returns number of variables.
}
