% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_covratio}
\alias{bt_covratio}
\title{Covariance Ratio}
\usage{
bt_covratio(model)
}
\arguments{
\item{model}{A fitted linear model.}
}
\value{
Numeric vector.
}
\description{
Calculates Covariance Ratios (COVRATIO).
}
\examples{
model <- lm(mpg ~ wt, data = mtcars)
bt_covratio(model)

}
