% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_crosstab}
\alias{bt_crosstab}
\title{Crosstab}
\usage{
bt_crosstab(data, row, column)
}
\arguments{
\item{data}{A data frame.}

\item{row}{Variable for rows.}

\item{column}{Variable for columns.}
}
\value{
Matrix.
}
\description{
Creates a contingency table between two categorical variables.
}
\examples{
birds <- data.frame(
  diet = c(
    "Insectivore",
    "Omnivore",
    "Carnivore",
    "Insectivore"
  ),
  status = c(
    "LC",
    "LC",
    "VU",
    "NT"
  ),
  stringsAsFactors = FALSE
)

bt_crosstab(
  birds,
  "diet",
  "status"
)

}
