% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bt_data_dictionary}
\alias{bt_data_dictionary}
\title{Data Dictionary}
\usage{
bt_data_dictionary(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Data frame.
}
\description{
Creates a simple data dictionary.
}
\examples{
birds <- data.frame(
  species = c("A","B"),
  mass = c(20,25),
  stringsAsFactors = FALSE
)

bt_data_dictionary(birds)

}
