% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bt_dataset_summary}
\alias{bt_dataset_summary}
\title{Bird Dataset Summary}
\usage{
bt_dataset_summary(
  data,
  species_col = "species",
  family_col = "family",
  order_col = "order"
)
}
\arguments{
\item{data}{A data frame.}

\item{species_col}{Species column.}

\item{family_col}{Family column.}

\item{order_col}{Order column.}
}
\value{
BirdTraitsSummary object.
}
\description{
Produces a concise summary of a bird dataset.
}
\examples{
birds <- data.frame(
  species=c("A","B","C","A"),
  family=c("F1","F1","F2","F1"),
  order=c("O1","O1","O2","O1"),
  stringsAsFactors=FALSE
)

bt_dataset_summary(birds)

}
