% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bt_date_columns}
\alias{bt_date_columns}
\title{Date Columns}
\usage{
bt_date_columns(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Character vector.
}
\description{
Returns the names of Date columns.
}
\examples{
birds <- data.frame(
  survey_date = as.Date(c("2024-01-01","2024-01-02"))
)

bt_date_columns(birds)

}
