% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_deviance}
\alias{bt_deviance}
\title{Model Deviance}
\usage{
bt_deviance(model)
}
\arguments{
\item{model}{A fitted model.}
}
\value{
Numeric.
}
\description{
Returns model deviance.
}
\examples{
birds <- data.frame(
  y = c(20,25,18,30,27),
  x = c(12,14,11,16,15)
)

model <- lm(y ~ x, data = birds)

bt_deviance(model)

}
