% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_df}
\alias{bt_df}
\title{Degrees of Freedom}
\usage{
bt_df(model)
}
\arguments{
\item{model}{A fitted model.}
}
\value{
Integer vector.
}
\description{
Returns model degrees of freedom.
}
\examples{
birds <- data.frame(
  y = c(20,25,18,30,27),
  x = c(12,14,11,16,15)
)

model <- lm(y ~ x, data = birds)

bt_df(model)

}
