% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_dfbetas}
\alias{bt_dfbetas}
\title{DFBETAs}
\usage{
bt_dfbetas(model)
}
\arguments{
\item{model}{A fitted regression model.}
}
\value{
Matrix.
}
\description{
Calculates DFBETAs for each coefficient.
}
\examples{
model <- lm(mpg ~ wt, data = mtcars)
bt_dfbetas(model)

}
