% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_dffits}
\alias{bt_dffits}
\title{DFFITS}
\usage{
bt_dffits(model)
}
\arguments{
\item{model}{A fitted regression model.}
}
\value{
Numeric vector.
}
\description{
Calculates DFFITS statistics.
}
\examples{
model <- lm(mpg ~ wt, data = mtcars)
bt_dffits(model)

}
