% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_dissimilarity_matrix}
\alias{bt_dissimilarity_matrix}
\title{Pairwise Dissimilarity Matrix}
\usage{
bt_dissimilarity_matrix(x, method = "bray")
}
\arguments{
\item{x}{Community matrix.}

\item{method}{Distance method.}
}
\value{
Matrix.
}
\description{
Converts a beta-diversity distance object into a matrix.
}
\examples{
mat <- matrix(
c(
5,2,0,
4,1,3,
0,5,2,
2,1,6
),
nrow = 4,
byrow = TRUE
)

bt_dissimilarity_matrix(mat)

}
