% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_diversity_summary}
\alias{bt_diversity_summary}
\title{Diversity Summary}
\usage{
bt_diversity_summary(
  data,
  species = "species",
  family = "family",
  order = "order"
)
}
\arguments{
\item{data}{Data frame.}

\item{species}{Species column.}

\item{family}{Family column.}

\item{order}{Order column.}
}
\value{
BirdTraitsDiversity object.
}
\description{
Summarizes species, families and orders.
}
