% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_duplicate_rows}
\alias{bt_duplicate_rows}
\title{Count Duplicate Rows}
\usage{
bt_duplicate_rows(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Integer.
}
\description{
Counts duplicated observations.
}
