% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_durbin_watson}
\alias{bt_durbin_watson}
\title{Check Autocorrelation}
\usage{
bt_durbin_watson(model)
}
\arguments{
\item{model}{A fitted linear model.}
}
\value{
Numeric.
}
\description{
Computes the Durbin-Watson statistic.
}
\details{
Values close to 2 indicate little evidence of
autocorrelation.
}
\examples{
birds <- data.frame(
  y = c(20,25,18,30,27,24),
  x = c(12,14,11,16,15,13)
)

model <- lm(y ~ x, data = birds)

bt_durbin_watson(model)

}
