% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_empty_strings}
\alias{bt_empty_strings}
\title{Count Empty Strings}
\usage{
bt_empty_strings(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Integer.
}
\description{
Counts empty text values.
}
