% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_evenness}
\alias{bt_evenness}
\title{Pielou Evenness Index}
\usage{
bt_evenness(data, species = "species")
}
\arguments{
\item{data}{Data frame.}

\item{species}{Species column.}
}
\value{
Numeric.
}
\description{
Calculates community evenness.
}
\examples{
birds <- data.frame(
  species=c(
    "A","A","A",
    "B","B",
    "C",
    "D"
  )
)

bt_evenness(birds)

}
