% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_exponentiate_coefficients}
\alias{bt_exponentiate_coefficients}
\title{Exponentiated Coefficients}
\usage{
bt_exponentiate_coefficients(model)
}
\arguments{
\item{model}{A fitted model.}
}
\value{
Numeric vector.
}
\description{
Exponentiated Coefficients
}
