% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{bt_export_report}
\alias{bt_export_report}
\title{Export Report as Text File}
\usage{
bt_export_report(report, filename)
}
\arguments{
\item{report}{Report object (from bt_create_report).}

\item{filename}{File path to save report.}
}
\value{
Invisibly returns the filename.
}
\description{
Saves a report to a text file.
}
\examples{
birds <- data.frame(
  species = c("A", "B", "C"),
  mass = c(20, 25, 30)
)

report <- bt_create_report(birds)
# bt_export_report(report, "bird_report.txt")

}
