% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bt_factor_columns}
\alias{bt_factor_columns}
\title{Factor Columns}
\usage{
bt_factor_columns(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Character vector.
}
\description{
Returns the names of factor columns.
}
\examples{
birds <- data.frame(
  status = factor(c("LC","VU"))
)

bt_factor_columns(birds)

}
