% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_family_richness}
\alias{bt_family_richness}
\title{Family Richness}
\usage{
bt_family_richness(data, family = "family")
}
\arguments{
\item{data}{Data frame.}

\item{family}{Family column.}
}
\value{
Integer.
}
\description{
Counts unique bird families.
}
