% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{bt_filter_family}
\alias{bt_filter_family}
\title{Filter by Family}
\usage{
bt_filter_family(data, family)
}
\arguments{
\item{data}{Data frame.}

\item{family}{Character vector.}
}
\value{
Filtered data frame.
}
\description{
Filter by Family
}
