% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{bt_filter_iucn}
\alias{bt_filter_iucn}
\title{Filter by IUCN Category}
\usage{
bt_filter_iucn(data, iucn)
}
\arguments{
\item{data}{Data frame.}

\item{iucn}{Character vector.}
}
\value{
Filtered data frame.
}
\description{
Filter by IUCN Category
}
