% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{bt_filter_multiple}
\alias{bt_filter_multiple}
\title{Filter Using Multiple Conditions}
\usage{
bt_filter_multiple(data, ...)
}
\arguments{
\item{data}{Data frame.}

\item{...}{Named filters, e.g. family="Accipitridae".}
}
\value{
Filtered data frame.
}
\description{
Filter Using Multiple Conditions
}
