% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{bt_filter_order}
\alias{bt_filter_order}
\title{Filter by Order}
\usage{
bt_filter_order(data, order)
}
\arguments{
\item{data}{Data frame.}

\item{order}{Character vector.}
}
\value{
Filtered data frame.
}
\description{
Filter by Order
}
