% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{bt_filter_species}
\alias{bt_filter_species}
\title{Filter by Species}
\usage{
bt_filter_species(data, species)
}
\arguments{
\item{data}{Data frame.}

\item{species}{Character vector of species names.}
}
\value{
Filtered data frame.
}
\description{
Filter by Species
}
