% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{bt_filter_trait}
\alias{bt_filter_trait}
\title{Filter by Any Trait}
\usage{
bt_filter_trait(data, trait, value)
}
\arguments{
\item{data}{Data frame.}

\item{trait}{Column name.}

\item{value}{Value(s) to keep.}
}
\value{
Filtered data frame.
}
\description{
Filter by Any Trait
}
