% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_fisher_alpha}
\alias{bt_fisher_alpha}
\title{Fisher's Alpha Diversity}
\usage{
bt_fisher_alpha(x)
}
\arguments{
\item{x}{Community matrix.}
}
\value{
Numeric.
}
\description{
Calculates Fisher's Alpha diversity index.
}
\details{
Wrapper around vegan::fisher.alpha().
}
\examples{
mat <- matrix(
c(
5,2,0,
4,1,3
),
nrow=2,
byrow=TRUE
)

bt_fisher_alpha(mat)

}
