% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_fisher_test}
\alias{bt_fisher_test}
\title{Fisher Exact Test}
\usage{
bt_fisher_test(data, row, column)
}
\arguments{
\item{data}{A data frame.}

\item{row}{Row variable.}

\item{column}{Column variable.}
}
\value{
htest object.
}
\description{
Performs Fisher's Exact Test.
}
