% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_fitted}
\alias{bt_fitted}
\title{Fitted Values}
\usage{
bt_fitted(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
Numeric vector.
}
\description{
Returns fitted values from a model.
}
\examples{
birds <- data.frame(
  y = c(20,25,18,30,27),
  x = c(12,14,11,16,15)
)

model <- lm(y ~ x, data = birds)

bt_fitted(model)

}
