% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bt_glimpse}
\alias{bt_glimpse}
\title{Glimpse Dataset}
\usage{
bt_glimpse(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Invisibly returns the data.
}
\description{
Displays a compact overview of a dataset.
}
\examples{
birds <- data.frame(
  species = c("A","B"),
  mass = c(20,25),
  stringsAsFactors = FALSE
)

bt_glimpse(birds)

}
