% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_incidence_rate_ratio}
\alias{bt_incidence_rate_ratio}
\title{Incidence Rate Ratios}
\usage{
bt_incidence_rate_ratio(model)
}
\arguments{
\item{model}{Poisson glm.}
}
\value{
Numeric vector.
}
\description{
Incidence Rate Ratios
}
