% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_jaccard}
\alias{bt_jaccard}
\title{Jaccard Distance}
\usage{
bt_jaccard(x, binary = TRUE)
}
\arguments{
\item{x}{Community matrix.}

\item{binary}{Treat data as presence/absence?}
}
\value{
A dist object.
}
\description{
Calculates pairwise Jaccard dissimilarity.
}
\details{
Wrapper around vegan::vegdist().
}
\examples{
mat <- matrix(
c(
5,2,0,
4,1,3,
0,5,2,
2,1,6
),
nrow = 4,
byrow = TRUE
)

bt_jaccard(mat)

}
