% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_kruskal_test}
\alias{bt_kruskal_test}
\title{Kruskal-Wallis Test}
\usage{
bt_kruskal_test(data, response, group)
}
\arguments{
\item{data}{A data frame.}

\item{response}{Numeric variable.}

\item{group}{Grouping variable.}
}
\value{
htest object.
}
\description{
Performs a Kruskal-Wallis test.
}
