% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_leverage}
\alias{bt_leverage}
\title{Leverage Values}
\usage{
bt_leverage(model)
}
\arguments{
\item{model}{A fitted regression model.}
}
\value{
Numeric vector.
}
\description{
Calculates leverage (hat values).
}
\examples{
model <- lm(mpg ~ wt, data = mtcars)
bt_leverage(model)

}
