% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_likelihood_ratio_test}
\alias{bt_likelihood_ratio_test}
\title{Likelihood Ratio Test}
\usage{
bt_likelihood_ratio_test(...)
}
\arguments{
\item{...}{Two or more fitted model objects.}
}
\value{
An ANOVA table.
}
\description{
Performs a likelihood ratio test for two or more nested models.
}
\details{
This is a wrapper around
\code{stats::anova(..., test = "Chisq")}.

Models should be nested for valid inference.
}
\examples{
birds <- data.frame(
  y = c(10,12,13,16,18),
  x1 = c(2,3,4,5,6),
  x2 = c(5,4,6,7,8)
)

m1 <- glm(
  y ~ x1,
  data = birds,
  family = poisson()
)

m2 <- glm(
  y ~ x1 + x2,
  data = birds,
  family = poisson()
)

bt_likelihood_ratio_test(m1, m2)

}
