% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_linear_regression}
\alias{bt_linear_regression}
\title{Linear Regression}
\usage{
bt_linear_regression(data, response, predictors)
}
\arguments{
\item{data}{A data frame.}

\item{response}{Response variable.}

\item{predictors}{Predictor variable(s).}
}
\value{
lm object.
}
\description{
Fits a linear regression model.
}
\examples{
birds <- data.frame(
  mass = c(20,25,18,30,27),
  wing = c(12,14,11,16,15)
)

bt_linear_regression(
  birds,
  "mass",
  "wing"
)

}
