% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bt_logical_columns}
\alias{bt_logical_columns}
\title{Logical Columns}
\usage{
bt_logical_columns(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Character vector.
}
\description{
Returns the names of logical columns.
}
\examples{
birds <- data.frame(
  migratory = c(TRUE,FALSE)
)

bt_logical_columns(birds)

}
