% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_logistic_regression}
\alias{bt_logistic_regression}
\title{Logistic Regression}
\usage{
bt_logistic_regression(data, response, predictors)
}
\arguments{
\item{data}{A data frame.}

\item{response}{Binary response variable.}

\item{predictors}{Predictor variable(s).}
}
\value{
glm object.
}
\description{
Fits a logistic regression model.
}
