% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_logLik}
\alias{bt_logLik}
\title{Log-Likelihood}
\usage{
bt_logLik(model)
}
\arguments{
\item{model}{A fitted model.}
}
\value{
Numeric.
}
\description{
Returns model log-likelihood.
}
\examples{
birds <- data.frame(
  y = c(20,25,18,30,27),
  x = c(12,14,11,16,15)
)

model <- lm(y ~ x, data = birds)

bt_logLik(model)

}
