% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_mean}
\alias{bt_mean}
\title{Mean}
\usage{
bt_mean(data, column)
}
\arguments{
\item{data}{A data frame.}

\item{column}{Numeric column.}
}
\value{
Numeric.
}
\description{
Calculates the mean of a numeric variable.
}
\examples{
birds <- data.frame(
  mass = c(20,25,18,30)
)

bt_mean(birds, "mass")

}
