% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_median}
\alias{bt_median}
\title{Median}
\usage{
bt_median(data, column)
}
\arguments{
\item{data}{A data frame.}

\item{column}{Numeric column.}
}
\value{
Numeric.
}
\description{
Calculates the median of a numeric variable.
}
\examples{
birds <- data.frame(
  mass = c(20,25,18,30)
)

bt_median(birds, "mass")

}
