% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_menhinick_index}
\alias{bt_menhinick_index}
\title{Menhinick Richness Index}
\usage{
bt_menhinick_index(data, species = "species")
}
\arguments{
\item{data}{Data frame.}

\item{species}{Species column.}
}
\value{
Numeric.
}
\description{
Calculates Menhinick richness.
}
\examples{
birds <- data.frame(
  species=c(
    "A","A","A",
    "B","B",
    "C",
    "D"
  )
)

bt_menhinick_index(birds)

}
