% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_model_summary}
\alias{bt_model_summary}
\title{Model Summary Table}
\usage{
bt_model_summary(...)
}
\arguments{
\item{...}{Two or more fitted model objects.}
}
\value{
A data frame.
}
\description{
Returns a summary table for one or more fitted models.
}
\details{
The summary includes:
\itemize{
\item AIC
\item BIC
\item Log-Likelihood
\item Deviance
\item Residual Degrees of Freedom
\item Number of Parameters
}
}
\examples{
birds <- data.frame(
  y = c(10,12,13,16,18),
  x1 = c(2,3,4,5,6),
  x2 = c(5,4,6,7,8)
)

m1 <- lm(y ~ x1, data = birds)
m2 <- lm(y ~ x1 + x2, data = birds)

bt_model_summary(m1, m2)

}
