% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bt_numeric_columns}
\alias{bt_numeric_columns}
\title{Numeric Columns}
\usage{
bt_numeric_columns(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Character vector.
}
\description{
Returns the names of numeric columns.
}
\examples{
birds <- data.frame(
  mass = c(20,25),
  wing = c(10,11)
)

bt_numeric_columns(birds)

}
