% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bt_numeric_summary}
\alias{bt_numeric_summary}
\title{Numeric Summary}
\usage{
bt_numeric_summary(data, column)
}
\arguments{
\item{data}{A data frame.}

\item{column}{Numeric column.}
}
\value{
Data frame.
}
\description{
Returns descriptive statistics for a numeric column.
}
\examples{
birds <- data.frame(
  mass = c(20,25,18,30,27)
)

bt_numeric_summary(birds, "mass")

}
