% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_occurrence_frequency}
\alias{bt_occurrence_frequency}
\title{Occurrence Frequency}
\usage{
bt_occurrence_frequency(data, species = "species", site = "site")
}
\arguments{
\item{data}{Data frame.}

\item{species}{Species column.}

\item{site}{Sampling site column.}
}
\value{
Data frame.
}
\description{
Calculates the percentage of sampling units in which each species occurs.
}
\examples{
birds <- data.frame(
  site=c(1,1,2,2,3),
  species=c("A","B","A","C","A")
)

bt_occurrence_frequency(
  birds,
  site="site"
)

}
