% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_order_richness}
\alias{bt_order_richness}
\title{Order Richness}
\usage{
bt_order_richness(data, order = "order")
}
\arguments{
\item{data}{Data frame.}

\item{order}{Order column.}
}
\value{
Integer.
}
\description{
Counts unique bird orders.
}
