% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_percentage_table}
\alias{bt_percentage_table}
\title{Percentage Table}
\usage{
bt_percentage_table(data, column)
}
\arguments{
\item{data}{A data frame.}

\item{column}{Column name.}
}
\value{
Data frame.
}
\description{
Returns percentages for a categorical variable.
}
\examples{
birds <- data.frame(
  diet = c(
    "Insectivore",
    "Omnivore",
    "Insectivore",
    "Carnivore"
  ),
  stringsAsFactors = FALSE
)

bt_percentage_table(birds, "diet")

}
