% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{bt_plot_boxplot}
\alias{bt_plot_boxplot}
\title{Plot Boxplot}
\usage{
bt_plot_boxplot(data, y, group = NULL, fill = "#2C7FB8", outlier_color = "red")
}
\arguments{
\item{data}{A data frame.}

\item{y}{Numeric column.}

\item{group}{Optional grouping variable.}

\item{fill}{Fill colour.}

\item{outlier_color}{Colour of outliers.}
}
\value{
A ggplot object.
}
\description{
Creates a publication-ready boxplot for a numeric variable.
}
\details{
If a grouping variable is supplied, separate boxplots are produced
for each group.
}
\examples{
birds <- data.frame(
  mass = c(
    20,22,18,25,30,
    27,24,23,21,26
  ),
  diet = c(
    "A","A","A","B","B",
    "B","C","C","C","C"
  )
)

p1 <- bt_plot_boxplot(
  birds,
  y = "mass"
)

p2 <- bt_plot_boxplot(
  birds,
  y = "mass",
  group = "diet"
)

print(p1)
print(p2)

}
