% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{bt_plot_histogram}
\alias{bt_plot_histogram}
\title{Plot Histogram}
\usage{
bt_plot_histogram(data, variable, bins = 20, fill = "#2C7FB8", color = "black")
}
\arguments{
\item{data}{A data frame.}

\item{variable}{Numeric column name.}

\item{bins}{Number of histogram bins.}

\item{fill}{Fill colour.}

\item{color}{Border colour.}
}
\value{
A ggplot object.
}
\description{
Creates a publication-ready histogram for a numeric variable.
}
\examples{
birds <- data.frame(
  mass = c(
    20,22,18,25,30,
    27,24,23,21,26
  )
)

p <- bt_plot_histogram(
  birds,
  "mass"
)

print(p)

}
