% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{bt_plot_pie}
\alias{bt_plot_pie}
\title{Plot Trait Pie Chart}
\usage{
bt_plot_pie(data, trait, show_percent = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{trait}{Character column name.}

\item{show_percent}{Logical.}
}
\value{
A ggplot object.
}
\description{
Creates a publication-ready pie chart for a categorical trait.
}
\examples{
birds <- data.frame(
  diet = c(
    "Insectivore",
    "Omnivore",
    "Carnivore",
    "Insectivore"
  )
)

p <- bt_plot_pie(birds,"diet")
print(p)

}
