% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{bt_plot_scatter}
\alias{bt_plot_scatter}
\title{Plot Scatter Plot}
\usage{
bt_plot_scatter(
  data,
  x,
  y,
  color = "#2C7FB8",
  point_size = 3,
  alpha = 0.8,
  regression = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Numeric x-axis variable.}

\item{y}{Numeric y-axis variable.}

\item{color}{Point colour.}

\item{point_size}{Size of points.}

\item{alpha}{Point transparency.}

\item{regression}{Logical. Add regression line?}
}
\value{
A ggplot object.
}
\description{
Creates a publication-ready scatter plot between two numeric variables.
}
\examples{
birds <- data.frame(
  wing = c(12,14,11,16,15,13),
  mass = c(20,25,18,30,27,22)
)

p <- bt_plot_scatter(
  birds,
  x = "wing",
  y = "mass"
)

print(p)

}
