% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_diversity.R
\name{bt_plot_trait_space}
\alias{bt_plot_trait_space}
\title{Plot Trait Space}
\usage{
bt_plot_trait_space(data, x, y, species = NULL, color = NULL)
}
\arguments{
\item{data}{Data frame with species as rows and traits as columns.}

\item{x}{Character. Name of trait for x-axis.}

\item{y}{Character. Name of trait for y-axis.}

\item{species}{Optional column name for species labels.}

\item{color}{Optional column name for coloring points.}
}
\value{
A ggplot object.
}
\description{
Creates a 2D visualization of species in trait space.
}
\examples{
birds <- data.frame(
  species = c("A", "B", "C", "D"),
  mass = c(20, 25, 30, 22),
  wing = c(12, 14, 16, 13)
)

bt_plot_trait_space(birds, "mass", "wing", "species")

}
