% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_poisson_regression}
\alias{bt_poisson_regression}
\title{Poisson Regression}
\usage{
bt_poisson_regression(data, response, predictors)
}
\arguments{
\item{data}{A data frame.}

\item{response}{Count response.}

\item{predictors}{Predictor variable(s).}
}
\value{
glm object.
}
\description{
Fits a Poisson regression model.
}
