% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_predict}
\alias{bt_predict}
\title{Predict Values}
\usage{
bt_predict(model, newdata = NULL)
}
\arguments{
\item{model}{A fitted model.}

\item{newdata}{Optional new data.}
}
\value{
Numeric vector.
}
\description{
Generates predictions from a fitted model.
}
