% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_presence_absence_matrix}
\alias{bt_presence_absence_matrix}
\title{Presence Absence Matrix}
\usage{
bt_presence_absence_matrix(data, species = "species", site = "site")
}
\arguments{
\item{data}{Data frame.}

\item{species}{Species column.}

\item{site}{Site column.}
}
\value{
Matrix.
}
\description{
Creates a species-by-site presence absence matrix.
}
\examples{
birds <- data.frame(
  site=c(1,1,2,2,3),
  species=c("A","B","A","C","A")
)

bt_presence_absence_matrix(
  birds,
  site="site"
)

}
