% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_prop_table}
\alias{bt_prop_table}
\title{Proportion Table}
\usage{
bt_prop_table(data, row, column, margin = 1)
}
\arguments{
\item{data}{A data frame.}

\item{row}{Row variable.}

\item{column}{Column variable.}

\item{margin}{Margin for proportions.}
}
\value{
Matrix.
}
\description{
Creates a proportional contingency table.
}
\examples{
birds <- data.frame(
  diet = c(
    "Insectivore",
    "Omnivore",
    "Carnivore",
    "Insectivore"
  ),
  status = c(
    "LC",
    "LC",
    "VU",
    "NT"
  ),
  stringsAsFactors = FALSE
)

bt_prop_table(
  birds,
  "diet",
  "status"
)

}
