% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_quantiles}
\alias{bt_quantiles}
\title{Quantiles}
\usage{
bt_quantiles(data, column)
}
\arguments{
\item{data}{A data frame.}

\item{column}{Numeric column.}
}
\value{
Numeric vector.
}
\description{
Returns quartiles for a numeric variable.
}
\examples{
birds <- data.frame(
  mass = c(20,25,18,30,27)
)

bt_quantiles(birds, "mass")

}
