% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_range}
\alias{bt_range}
\title{Range}
\usage{
bt_range(data, column)
}
\arguments{
\item{data}{A data frame.}

\item{column}{Numeric column.}
}
\value{
Numeric vector.
}
\description{
Returns minimum and maximum values.
}
\examples{
birds <- data.frame(
  mass = c(20,25,18,30)
)

bt_range(birds, "mass")

}
