% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_rank_abundance}
\alias{bt_rank_abundance}
\title{Rank Abundance Plot}
\usage{
bt_rank_abundance(
  data,
  species = "species",
  abundance = NULL,
  color = "#2C7FB8",
  point_size = 3,
  line_size = 0.8,
  log_scale = FALSE
)
}
\arguments{
\item{data}{Data frame.}

\item{species}{Species column.}

\item{abundance}{Optional abundance column.
If NULL, abundance is calculated from species frequencies.}

\item{color}{Line and point colour.}

\item{point_size}{Size of points.}

\item{line_size}{Width of connecting line.}

\item{log_scale}{Logical. Plot abundance on log10 scale?}
}
\value{
A ggplot object.
}
\description{
Creates a Whittaker Rank-Abundance plot showing species ranked from
most abundant to least abundant.
}
\examples{
birds <- data.frame(
  species = c(
    "A","A","A",
    "B","B",
    "C",
    "D","D","D","D"
  )
)

p <- bt_rank_abundance(birds)
print(p)

}
