% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_rank_abundance_table}
\alias{bt_rank_abundance_table}
\title{Rank-Abundance Table}
\usage{
bt_rank_abundance_table(data, species = "species", abundance = NULL)
}
\arguments{
\item{data}{Data frame.}

\item{species}{Species column.}

\item{abundance}{Optional abundance column.}
}
\value{
Data frame.
}
\description{
Returns the ranked abundance table used to create the Whittaker plot.
}
