% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_rank_models}
\alias{bt_rank_models}
\title{Rank Models}
\usage{
bt_rank_models(...)
}
\arguments{
\item{...}{Two or more fitted model objects.}
}
\value{
A data frame.
}
\description{
Ranks multiple fitted models using Akaike Information Criterion (AIC).
}
\details{
The returned table contains:
\itemize{
\item AIC
\item Delta AIC
\item Akaike Weight
\item Rank
}

Models are ordered from best (lowest AIC) to worst.
}
\examples{
birds <- data.frame(
  y = c(10,12,13,16,18),
  x1 = c(2,3,4,5,6),
  x2 = c(5,4,6,7,8)
)

m1 <- lm(y ~ x1, data = birds)
m2 <- lm(y ~ x1 + x2, data = birds)

bt_rank_models(m1, m2)

}
