% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_rarefaction}
\alias{bt_rarefaction}
\title{Rarefaction}
\usage{
bt_rarefaction(x, sample)
}
\arguments{
\item{x}{Community matrix.}

\item{sample}{Sample size.}
}
\value{
Numeric vector.
}
\description{
Calculates expected species richness for a standardized sample size.
}
\details{
Wrapper around vegan::rarefy().
}
\examples{
mat <- matrix(
c(
5,2,0,
4,1,3
),
nrow=2,
byrow=TRUE
)

bt_rarefaction(
mat,
sample = 5
)

}
