% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_relative_abundance}
\alias{bt_relative_abundance}
\title{Relative Abundance}
\usage{
bt_relative_abundance(data, species = "species")
}
\arguments{
\item{data}{Data frame.}

\item{species}{Species column.}
}
\value{
Data frame.
}
\description{
Calculates abundance and relative abundance (\%) of each species.
}
\examples{
birds <- data.frame(
  species=c("A","A","B","C","C","C")
)

bt_relative_abundance(birds)

}
