% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{bt_remove_duplicates}
\alias{bt_remove_duplicates}
\title{Remove Duplicate Rows}
\usage{
bt_remove_duplicates(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Data frame.
}
\description{
Remove Duplicate Rows
}
