% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{bt_remove_empty}
\alias{bt_remove_empty}
\title{Remove Empty Rows}
\usage{
bt_remove_empty(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Cleaned data frame.
}
\description{
Removes rows containing only missing values.
}
