% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_residuals}
\alias{bt_residuals}
\title{Model Residuals}
\usage{
bt_residuals(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
Numeric vector.
}
\description{
Returns residuals from a fitted model.
}
\examples{
birds <- data.frame(
  y = c(20,25,18,30,27),
  x = c(12,14,11,16,15)
)

model <- lm(y ~ x, data = birds)

bt_residuals(model)

}
