% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_rse}
\alias{bt_rse}
\title{Residual Standard Error}
\usage{
bt_rse(model)
}
\arguments{
\item{model}{A linear model.}
}
\value{
Numeric.
}
\description{
Returns the residual standard error of a linear model.
}
\examples{
birds <- data.frame(
  y = c(20,25,18,30,27),
  x = c(12,14,11,16,15)
)

model <- lm(y ~ x, data = birds)

bt_rse(model)

}
