% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{bt_select_columns}
\alias{bt_select_columns}
\title{Select Columns}
\usage{
bt_select_columns(data, columns)
}
\arguments{
\item{data}{Data frame.}

\item{columns}{Character vector.}
}
\value{
Data frame.
}
\description{
Select Columns
}
