% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_shannon_index}
\alias{bt_shannon_index}
\title{Shannon Diversity Index}
\usage{
bt_shannon_index(data, species = "species")
}
\arguments{
\item{data}{Data frame.}

\item{species}{Species column.}
}
\value{
Numeric.
}
\description{
Calculates Shannon-Wiener diversity index.
}
\examples{
birds <- data.frame(
  species=c(
    "A","A","A",
    "B","B",
    "C",
    "D"
  )
)

bt_shannon_index(birds)

}
