% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_simpson_index}
\alias{bt_simpson_index}
\title{Simpson Diversity Index}
\usage{
bt_simpson_index(data, species = "species")
}
\arguments{
\item{data}{Data frame.}

\item{species}{Species column.}
}
\value{
Numeric.
}
\description{
Calculates Simpson diversity (1-D).
}
\examples{
birds <- data.frame(
  species=c(
    "A","A","A",
    "B","B",
    "C",
    "D"
  )
)

bt_simpson_index(birds)

}
