% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_sorensen}
\alias{bt_sorensen}
\title{Sorensen Distance}
\usage{
bt_sorensen(x)
}
\arguments{
\item{x}{Community matrix.}
}
\value{
A dist object.
}
\description{
Calculates Sorensen (Dice) dissimilarity.
}
\details{
In vegan this is obtained using
Jaccard distance with binary = TRUE.
}
\examples{
mat <- matrix(
c(
5,2,0,
4,1,3,
0,5,2,
2,1,6
),
nrow = 4,
byrow = TRUE
)

bt_sorensen(mat)

}
