% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_species_accumulation}
\alias{bt_species_accumulation}
\title{Species Accumulation}
\usage{
bt_species_accumulation(x, method = "random")
}
\arguments{
\item{x}{Community matrix.}

\item{method}{Accumulation method.}
}
\value{
specaccum object.
}
\description{
Computes species accumulation curves.
}
\details{
Wrapper around vegan::specaccum().
}
\examples{
mat <- matrix(
c(
5,2,0,
4,1,3,
1,0,7,
6,1,2
),
nrow=4,
byrow=TRUE
)

bt_species_accumulation(mat)

}
