% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_species_frequency}
\alias{bt_species_frequency}
\title{Species Frequency}
\usage{
bt_species_frequency(data, species = "species")
}
\arguments{
\item{data}{Data frame.}

\item{species}{Species column.}
}
\value{
Data frame.
}
\description{
Counts the number of records for each species.
}
\examples{
birds <- data.frame(
  species=c("A","A","B","C")
)

bt_species_frequency(birds)

}
