% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{bt_species_richness}
\alias{bt_species_richness}
\title{Species Richness}
\usage{
bt_species_richness(data, species = "species")
}
\arguments{
\item{data}{Data frame.}

\item{species}{Column containing species names.}
}
\value{
Integer.
}
\description{
Calculates the number of unique species.
}
\examples{
birds <- data.frame(
  species = c("A","B","C","A")
)

bt_species_richness(birds)

}
