% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_standard_deviation}
\alias{bt_standard_deviation}
\title{Standard Deviation}
\usage{
bt_standard_deviation(data, column)
}
\arguments{
\item{data}{A data frame.}

\item{column}{Numeric column.}
}
\value{
Numeric.
}
\description{
Calculates standard deviation.
}
\examples{
birds <- data.frame(
  mass = c(20,25,18,30)
)

bt_standard_deviation(birds, "mass")

}
