% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{bt_standardize_missing}
\alias{bt_standardize_missing}
\title{Replace Missing Values}
\usage{
bt_standardize_missing(data, value = "Unknown")
}
\arguments{
\item{data}{Data frame.}

\item{value}{Replacement value.}
}
\value{
Cleaned data frame.
}
\description{
Replaces missing values in character columns.
}
