% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_standardized_coefficients}
\alias{bt_standardized_coefficients}
\title{Standardized Coefficients}
\usage{
bt_standardized_coefficients(model)
}
\arguments{
\item{model}{A fitted lm object.}
}
\value{
Named numeric vector.
}
\description{
Standardized Coefficients
}
