% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_statistics}
\alias{bt_statistics}
\title{Dataset Statistics}
\usage{
bt_statistics(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
BirdTraitsStatistics object.
}
\description{
Returns complete descriptive statistics.
}
\examples{
birds <- data.frame(
  species = c("A","B","C","A"),
  family = c("F1","F1","F2","F1"),
  order = c("O1","O1","O2","O1")
)

bt_statistics(birds)

}
