% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bt_summary_report}
\alias{bt_summary_report}
\title{Summary Report}
\usage{
bt_summary_report(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Invisibly returns TRUE.
}
\description{
Produces a complete summary report.
}
\examples{
birds <- data.frame(
  species=c("A","B"),
  family=c("F1","F1"),
  order=c("O1","O1"),
  stringsAsFactors=FALSE
)

bt_summary_report(birds)

}
