% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_t_test}
\alias{bt_t_test}
\title{Two Sample t-test}
\usage{
bt_t_test(data, response, group)
}
\arguments{
\item{data}{A data frame.}

\item{response}{Numeric variable.}

\item{group}{Grouping variable.}
}
\value{
htest object.
}
\description{
Performs an independent t-test.
}
\examples{
birds <- data.frame(
  mass=c(20,25,18,27),
  sex=c("M","M","F","F")
)

bt_t_test(
  birds,
  "mass",
  "sex"
)

}
