% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservation.R
\name{bt_threatened_species}
\alias{bt_threatened_species}
\title{Identify Threatened Species}
\usage{
bt_threatened_species(data, species = "species", iucn = "iucn_status")
}
\arguments{
\item{data}{Data frame.}

\item{species}{Column name containing species names.}

\item{iucn}{Column name containing IUCN status.}
}
\value{
Data frame with threatened species and their status.
}
\description{
Returns list of species categorized as threatened (CR, EN, or VU).
}
\examples{
birds <- data.frame(
  species = c("A", "B", "C", "D", "E"),
  iucn = c("LC", "VU", "EN", "LC", "CR")
)

bt_threatened_species(birds, "species", "iucn")

}
