% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{bt_trait_frequency}
\alias{bt_trait_frequency}
\title{Trait Frequency}
\usage{
bt_trait_frequency(data, trait)
}
\arguments{
\item{data}{Data frame.}

\item{trait}{Trait column.}
}
\value{
Data frame.
}
\description{
Returns only frequency counts.
}
