% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{bt_trait_levels}
\alias{bt_trait_levels}
\title{Trait Levels}
\usage{
bt_trait_levels(data, trait)
}
\arguments{
\item{data}{Data frame.}

\item{trait}{Trait column.}
}
\value{
Character vector.
}
\description{
Lists unique levels of a trait.
}
