% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{bt_trait_missing}
\alias{bt_trait_missing}
\title{Missing Trait Values}
\usage{
bt_trait_missing(data, trait)
}
\arguments{
\item{data}{Data frame.}

\item{trait}{Trait column.}
}
\value{
Integer.
}
\description{
Counts missing values.
}
