% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{bt_trait_percentage}
\alias{bt_trait_percentage}
\title{Trait Percentage}
\usage{
bt_trait_percentage(data, trait)
}
\arguments{
\item{data}{Data frame.}

\item{trait}{Trait column.}
}
\value{
Data frame.
}
\description{
Returns percentages only.
}
