% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{bt_trait_summary}
\alias{bt_trait_summary}
\title{Summarize a Bird Trait}
\usage{
bt_trait_summary(data, trait)
}
\arguments{
\item{data}{A data frame.}

\item{trait}{Column name containing the trait.}
}
\value{
A BirdTraitsTrait object.
}
\description{
Produces counts and percentages for any categorical trait.
}
\examples{
birds <- data.frame(
  species = c("A", "B", "C"),
  diet = c("Insectivore", "Omnivore", "Carnivore"),
  stringsAsFactors = FALSE
)

bt_trait_summary(birds, "diet")

}
