% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bt_variable_types}
\alias{bt_variable_types}
\title{Variable Types}
\usage{
bt_variable_types(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Data frame.
}
\description{
Counts variables by data type.
}
\examples{
birds <- data.frame(
  species = c("A","B"),
  mass = c(20,25),
  status = factor(c("LC","VU")),
  migratory = c(TRUE,FALSE),
  survey_date = as.Date(c("2024-01-01","2024-01-02")),
  stringsAsFactors = FALSE
)

bt_variable_types(birds)

}
