% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_variance}
\alias{bt_variance}
\title{Variance}
\usage{
bt_variance(data, column)
}
\arguments{
\item{data}{A data frame.}

\item{column}{Numeric column.}
}
\value{
Numeric.
}
\description{
Calculates variance.
}
\examples{
birds <- data.frame(
  mass = c(20,25,18,30)
)

bt_variance(birds, "mass")

}
