% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bt_wilcox_test}
\alias{bt_wilcox_test}
\title{Wilcoxon Rank Sum Test}
\usage{
bt_wilcox_test(data, response, group)
}
\arguments{
\item{data}{A data frame.}

\item{response}{Numeric variable.}

\item{group}{Grouping variable.}
}
\value{
htest object.
}
\description{
Performs a Wilcoxon test.
}
