% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.BirdTraitsConservation}
\alias{print.BirdTraitsConservation}
\title{Print Method for BirdTraitsConservation}
\usage{
\method{print}{BirdTraitsConservation}(x, ...)
}
\arguments{
\item{x}{A BirdTraitsConservation data frame.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns x.
}
\description{
Custom print method for BirdTraitsConservation objects.
}
