% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.BirdTraitsValidation.R
\name{print.BirdTraitsValidation}
\alias{print.BirdTraitsValidation}
\title{Print Validation Summary}
\usage{
\method{print}{BirdTraitsValidation}(x, ...)
}
\arguments{
\item{x}{Validation object}

\item{...}{ignored}
}
\description{
Print Validation Summary
}
