% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.BirdTraitsConservation}
\alias{summary.BirdTraitsConservation}
\title{Summary Method for BirdTraitsConservation Objects}
\usage{
\method{summary}{BirdTraitsConservation}(object, ...)
}
\arguments{
\item{object}{A BirdTraitsConservation object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns object.
}
\description{
Summary Method for BirdTraitsConservation Objects
}
