% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.BirdTraitsDiversity}
\alias{summary.BirdTraitsDiversity}
\title{Summary Method for BirdTraitsDiversity Objects}
\usage{
\method{summary}{BirdTraitsDiversity}(object, ...)
}
\arguments{
\item{object}{A BirdTraitsDiversity object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns object.
}
\description{
Summary Method for BirdTraitsDiversity Objects
}
