% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.BirdTraitsReport}
\alias{summary.BirdTraitsReport}
\title{Summary Method for BirdTraitsReport Objects}
\usage{
\method{summary}{BirdTraitsReport}(object, ...)
}
\arguments{
\item{object}{A BirdTraitsReport object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns object.
}
\description{
Summary Method for BirdTraitsReport Objects
}
